backend/auth-server/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
firebase-admin==6.1.0
python-dotenv==1.0.0
requests==2.31.0
